feat: extract enforce call logic into custom hook useEnforceCall
